requirements-dev.txt
pytest
pytest-aiohttp
ruff
mypy
bandit
